requirements.txt
firebase-admin>=7.0.0
fastapi>=0.116.1
uvicorn>=0.35.0
email-validator>=2.1.0
requests>=2.31.0
google-auth>=2.23.4
google-auth-oauthlib>=1.1.0